pyproject.toml
[project]
name = "favboard"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "gspread>=6.2.1",
    "python-dotenv>=1.2.1",
    "streamlit",
]
